requirements.txt
absl-py==2.3.1
accelerate==0.32.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
contextlib2==21.6.0
einops==0.6.1
exceptiongroup==1.3.0
fastapi==0.121.0
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.1
idna==3.11
Jinja2==3.1.6
julius==0.2.7
MarkupSafe==3.0.3
ml_collections==0.1.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
psutil==7.1.3
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
python-multipart==0.0.20
PyYAML==6.0.3
safetensors==0.6.2
-e git+https://github.com/jksikora/SCNet.git#egg=scnet
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
starlette==0.49.3
sympy==1.14.0
torch==2.9.0
torchaudio==2.9.0
torchcodec==0.8.1
tqdm==4.67.1
triton==3.5.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
zipstream-ng==1.9.0
